Parser in work

+++ BottApp.Database/User/IUserRepository.cs
@@ -9,7 +9,7 @@ namespace BottApp.Database.User
 
         Task<UserModel> GetOneByUid(long uid);
 
-        Task<UserModel?> FindOneByUid(int userId);
+        Task<UserModel?> FindOneByUid(long userId);
 
         Task<bool> UpdateUserPhone(UserModel model, string phone);
